#441 Update GraphQL folder structure
